fixed menu animation

+++ fvwm/menus.c
@@ -1077,7 +1077,7 @@ int DoMenusOverlap(MenuRoot *mr, int x, int y, int width, int height,
   x_overlap = 0;
   if (fTolerant) {
     /* Don't use multiplier if doing an intolerant check */
-    prior_width *= mr->ms->feel.PopupOffsetPercent / 100;
+    prior_width *= (float)(mr->ms->feel.PopupOffsetPercent) / 100.0;
   }
   if (y <= prior_y + prior_height - tolerance2 &&
       prior_y <= y + height - tolerance2 &&
@@ -1104,8 +1104,8 @@ int DoMenusOverlap(MenuRoot *mr, int x, int y, int width, int height,
  *
  ***********************************************************************/
 static
-Bool FPopupMenu (MenuRoot *menu, MenuRoot *menuPrior, int x, int y,
-		 Bool fWarpItem, MenuOptions *pops, Bool *ret_overlap)
+Bool FPopupMenu(MenuRoot *menu, MenuRoot *menuPrior, int x, int y,
+		Bool fWarpItem, MenuOptions *pops, Bool *ret_overlap)
 {
   Bool fWarpTitle = FALSE;
   int x_overlap, x_clipped_overlap;
